# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk mencatat servis kendaraan dan manajemen foto per transaksi.",
        "title": "Service Kendaraan API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/transaction": {
            "get": {
                "description": "Mengambil data bengkel.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Ambil semua list transaksi",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Halaman ke berapa (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Jumlah data per halaman (bisa 5, 10, 15, dll)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "tanggal_transaksi",
                        "description": "Kolom untuk sorting (tanggal_transaksi, created_at)",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "desc",
                        "description": "Arah urutan (asc, desc)",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Transaction"
                        }
                    }
                }
            },
            "post": {
                "description": "Simpan header transaksi beserta banyak detail service dan sparepart sekaligus.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Simpan Transaksi Lengkap",
                "parameters": [
                    {
                        "type": "string",
                        "description": "No Polisi Kendaraan",
                        "name": "no_polisi",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID Bengkel",
                        "name": "bengkel_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Total Biaya (Bruto)",
                        "name": "biaya_bruto",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Total Diskon",
                        "name": "diskon",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Total Bayar (Netto)",
                        "name": "biaya_netto",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Format: 2006-01-02",
                        "name": "tanggal_transaksi",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Format: 2006-01-02",
                        "name": "tanggal_masuk",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Format: 2006-01-02",
                        "name": "tanggal_selesai",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Kilometer Kendaraan",
                        "name": "km",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Nomor Nota Manual",
                        "name": "no_nota",
                        "in": "formData"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "file"
                        },
                        "collectionFormat": "multi",
                        "description": "Foto Nota Utama (Bisa upload lebih dari 1 foto)",
                        "name": "foto_nota[]",
                        "in": "formData"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "Daftar Nama Service",
                        "name": "service_name[]",
                        "in": "formData"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "number"
                        },
                        "collectionFormat": "multi",
                        "description": "Daftar Harga Service",
                        "name": "service_price[]",
                        "in": "formData"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "file"
                        },
                        "collectionFormat": "multi",
                        "description": "Daftar Foto Service",
                        "name": "foto_service[]",
                        "in": "formData"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "Daftar Nama Sparepart",
                        "name": "sparepart_name[]",
                        "in": "formData"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "Daftar Kode Sparepart",
                        "name": "sparepart_code[]",
                        "in": "formData"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "number"
                        },
                        "collectionFormat": "multi",
                        "description": "Daftar Harga Sparepart",
                        "name": "sparepart_price[]",
                        "in": "formData"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "Daftar Batas KM Sparepart",
                        "name": "sparepart_batas_km[]",
                        "in": "formData"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "Daftar Batas Waktu Sparepart",
                        "name": "sparepart_batas_waktu[]",
                        "in": "formData"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "boolean"
                        },
                        "collectionFormat": "multi",
                        "description": "Daftar Remind Sparepart",
                        "name": "sparepart_remind[]",
                        "in": "formData"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "number"
                        },
                        "collectionFormat": "multi",
                        "description": "Daftar Qty Sparepart",
                        "name": "sparepart_qty[]",
                        "in": "formData"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "file"
                        },
                        "collectionFormat": "multi",
                        "description": "Daftar Foto Sparepart",
                        "name": "foto_sparepart[]",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Transaction"
                        }
                    }
                }
            }
        },
        "/api/transaction/{id}": {
            "get": {
                "description": "Mengambil data transaksi lengkap termasuk riwayat service dan sparepartnya menggunakan ID dari URL",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Ambil detail satu transaksi berdasarkan ID beserta semua detailnya",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Transaksi",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data transaksi ditemukan",
                        "schema": {
                            "$ref": "#/definitions/models.Transaction"
                        }
                    },
                    "404": {
                        "description": "Data transaksi tidak ditemukan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Menghapus header transaksi, detail service, detail sparepart, serta menyapu bersih folder foto terkait di uploads.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Hapus Transaksi Bersih (Clean Delete)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Transaksi",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaksi dan seluruh aset berhasil dihapus",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Transaksi tidak ditemukan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.Bengkel": {
            "type": "object",
            "properties": {
                "aktif": {
                    "type": "boolean"
                },
                "alamat": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "jenis_service": {
                    "type": "string"
                },
                "kecamatan": {
                    "type": "string"
                },
                "kelurahan": {
                    "type": "string"
                },
                "kota_kab": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                },
                "no_telp": {
                    "type": "string"
                },
                "provinsi": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "example": "MOBIL BBM"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.DetailService": {
            "type": "object",
            "properties": {
                "biaya": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "foto_service": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nama_service": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.DetailSparepart": {
            "type": "object",
            "properties": {
                "batas_km": {
                    "type": "string"
                },
                "batas_waktu": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "foto_sparepart": {
                    "type": "string"
                },
                "harga_satuan": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "kode_sparepart": {
                    "type": "string"
                },
                "nama_sparepart": {
                    "type": "string"
                },
                "qty": {
                    "type": "number"
                },
                "remind": {
                    "type": "boolean"
                },
                "transaction_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Kendaraan": {
            "type": "object",
            "properties": {
                "aktif": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "jenis": {
                    "type": "string"
                },
                "manufaktur": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                },
                "no_polisi": {
                    "type": "string"
                },
                "tahun": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "warna": {
                    "type": "string"
                }
            }
        },
        "models.Transaction": {
            "type": "object",
            "properties": {
                "Bengkel": {
                    "$ref": "#/definitions/models.Bengkel"
                },
                "bengkel_id": {
                    "type": "integer"
                },
                "biaya_bruto": {
                    "type": "number"
                },
                "biaya_netto": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "details_service": {
                    "description": "Relasi",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.DetailService"
                    }
                },
                "details_sparepart": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.DetailSparepart"
                    }
                },
                "diskon": {
                    "type": "number"
                },
                "foto_nota": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "kendaraan": {
                    "$ref": "#/definitions/models.Kendaraan"
                },
                "km": {
                    "type": "string"
                },
                "no_nota": {
                    "type": "string"
                },
                "no_polisi": {
                    "type": "string"
                },
                "ppn": {
                    "type": "number"
                },
                "tanggal_masuk": {
                    "type": "string"
                },
                "tanggal_selesai": {
                    "type": "string"
                },
                "tanggal_transaksi": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    }
}